Switch to pyproject.toml configuration
Switches to `pyproject.toml` based configuration which is the preferred configuration for metadata as per
[PEP621](https://peps.python.org/pep-0621/).

Left a bunch of configuration in there for [ruff](https://github.com/astral-sh/ruff) which can be enabled/applied easily
using a [pre-commit](https://pre-commit.com) hook if desirable. This includes [black](https://github.com/psf/black)
formatting which I realise isn't to every one's taste.

Happy to add a `.pre-commit-config.yaml` if this is something acceptable.